# MAIN_project/static/openapi-schema.yml
openapi: 3.0.2
info:
  title: 'Maximum'
  version: ''
paths:
...
  /roles/parent-child/:
    get:
      security:
        - bearerAuth: [ ]
      summary: Get parent-child's list by token
      operationId: ParentChild
      description: |
        # Получение списка родителей-детей:
        ## Суперпользователь - абсолютно все записи "родитель-ребенок"
      parameters:
        - in: query
          name: ordering
          schema:
            type: string
          description: Order by id, parent, child
        - in: query
          name: limit
          schema:
            type: integer
          description: The numbers of items to return
        - in: query
          name: offset
          schema:
            type: integer
          description: The number of items to skip before starting to collect the result set
#        - in: query
#          name: page
#          schema:
#            type: integer
#          description: The page of items
      responses:
        '200':
          content:
            application/json:
              schema: { }
          description: OK
      tags:
        - roles

    post:
      security:
        - bearerAuth: [ ]
      operationId: CreateParentChild
      summary: Create a parent-child
      description: |
        # Создание роли родитель-ребенок
        ## Cуперпользователь - полный доступ
        ## Активный завуч - создание роли для активных пользователей с указанием любых ролей "родитель" и "ученик". Завуч должен быть в той же активной школе.
        Важно! При создании записи идет проверка на совпадение школы у родителя и ребенка (ученика).
      parameters: [ ]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                child_id:
                  type: integer
                  example: 1
                parent_id:
                  type: integer
                  example: 1
      responses:
        '201':
          content:
            application/json:
              schema: { }
          description: OK
      tags:
        - roles
...
  /objects/:
    get:
      security:
        - bearerAuth: [ ]
      summary: Get object's list by token
      operationId: Objects
      description: |
        # Получение списка предметов:
        ## Суперпользователь - абсолютно все предметы
        ## Активный пользователь - просмотр всех предметов в активных школах, в которых состоит и request.user
      parameters:
        - in: query
          name: school_id
          schema:
            type: integer
          description: filter by school_id
        - in: query
          name: search
          schema:
            type: string
          description: Search by name
        - in: query
          name: ordering
          schema:
            type: string
          description: Order by id, name, school
        - in: query
          name: limit
          schema:
            type: integer
          description: The numbers of items to return
        - in: query
          name: offset
          schema:
            type: integer
          description: The number of items to skip before starting to collect the result set
      responses:
        '200':
          content:
            application/json:
              schema: { }
          description: OK
      tags:
        - objects

    post:
      security:
        - bearerAuth: [ ]
      operationId: CreateObject
      summary: Create a object
      description: |
        # Создание предмета
        ## Cуперпользователь - создание предметов для любых школ
        ## Активный завуч - создание предметов только в своей активной школе
        ## В request body перечислены все возможные поля (все обязательные)
        ### Обязательные поля: "name", "school_id"
        ### Проверка на валидацию school_id
        ### Проверка на уникальность названия внутри школы
      parameters: [ ]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: name
                school_id:
                  type: integer
                  example: 1
      responses:
        '201':
          content:
            application/json:
              schema: { }
          description: OK
      tags:
        - objects
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    basicAuth:
      type: http
      scheme: basic